carros-eletricos-2/requirements.txt
flask==2.0.1
paho-mqtt==1.6.1
networkx==2.8.8
requests==2.28.1
pyyaml==6.0
werkzeug==2.2.3
pytest==7.4.0
pytest-mock==3.10.0
pytest-asyncio==0.21.0